git: verify-pack
